+++ CHANGELOG.md
@@ -9,4 +9,5 @@
 * post_processor/DockerSave: add support for packer's docker save post-processor
 * post_processor/DockerTag: add support for packer's docker tag post-processor
 * post_processor/GoogleComputeExport: add support for packer's google compute image exporter post-processor
+* post_processor/Manifest: add support for packer's manifest post-processor
 * provisioner/file: add support for packer's file provisioner